+++ src/plugins/intel_cpu/src/emitters/plugin/x64/jit_load_store_emitters.cpp
@@ -1303,7 +1303,7 @@ void jit_store_emitter::store_dword_to_word_extension(const Xbyak::Reg64& reg,
             if (store_num == 16) {
                 h->vmovdqu16(ptr[reg + offset], ymm);
             } else {
-                data_idx = static_cast<int>(ymm.getIdx());
+                data_idx = ymm.getIdx();
                 store_bytes<Vmm>(reg, offset, store_num * 2);
             }
         } else {
@@ -1324,7 +1324,7 @@ void jit_store_emitter::store_dword_to_word_extension(const Xbyak::Reg64& reg,
                 uni_vcvtneps2bf16_->emit_code({static_cast<size_t>(vmm.getIdx())}, {static_cast<size_t>(xmm.getIdx())});
             }
 
-            data_idx = static_cast<int>(xmm.getIdx());
+            data_idx = xmm.getIdx();
             store_bytes<Vmm>(reg, offset, store_num * 2);
         }
     } else if (is_f16) {
@@ -1339,7 +1339,7 @@ void jit_store_emitter::store_dword_to_word_extension(const Xbyak::Reg64& reg,
             if (store_num == 16) {
                 h->vmovdqu16(ptr[reg + offset], ymm);
             } else {
-                data_idx = static_cast<int>(ymm.getIdx());
+                data_idx = ymm.getIdx();
                 store_bytes<Vmm>(reg, offset, store_num * 2);
             }
         } else if (mayiuse(cpu::x64::avx2)) {
@@ -1351,7 +1351,7 @@ void jit_store_emitter::store_dword_to_word_extension(const Xbyak::Reg64& reg,
             if (store_num == 8) {
                 h->uni_vmovdqu16(ptr[reg + offset], xmm);
             } else {
-                data_idx = static_cast<int>(xmm.getIdx());
+                data_idx = xmm.getIdx();
                 store_bytes<Vmm>(reg, offset, store_num * 2);
             }
         } else {
